+++ R/snapshot.R
@@ -555,6 +555,9 @@ renv_snapshot_validate_dependencies_compatible <- function(project, lockfile, li
   packages <- extract_chr(records, "Package")
   locs <- find.package(packages, lib.loc = libpaths, quiet = TRUE)
   deps <- bapply(locs, renv_dependencies_discover_description)
+
+  # include constraints declared in the project's own DESCRIPTION
+  deps <- bind(list(deps, renv_project_requirements(project)))
   if (empty(deps))
     return(character())